backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
pandas
openpyxl
python-multipart
pillow
